plugin: Always install console_scripts of all system site packages
As of now, this plugin only handles dependencies specified in the
tox configuration(`deps=`), but package can have its own dependencies
or tox environment can have 'extras' requirement.

There is no direct API with the help of which it would be possible to
calculate deps of sdist or deps of project at develop mode.

Signed-off-by: Stanislav Levin <slev@altlinux.org>